Make TimerTask::tick return indication that method needs to be called
It could be that the scheduled method needs to accept an owner
of the TimerTask as a parameter. In that case we would get into circular
reference, as we would need to either pass the owner to the tick,
or to store reference to timer_task. None of these is possible without
penalty becase: 1) we already borrowed timer_task as mutable from the
owner, 2) circular reference is created.

This way, we have an option just to tell the owner that it needs
to call the method of its will.